+++ src/HandlerStack/ErrorHandlerStack.php
@@ -25,7 +25,7 @@ class ErrorHandlerStack extends AbstractHandlerStack
      */
     public function push($handler)
     {
-        $this->isolator()->set_error_handler($handler);
+        $this->isolator->set_error_handler($handler);
     }
 
     /**
@@ -35,12 +35,13 @@ class ErrorHandlerStack extends AbstractHandlerStack
      */
     public function pop()
     {
-        $handler = $this->isolator()->set_error_handler(function () {});
+        $handler = $this->isolator->set_error_handler(function () {});
+
         if (null !== $handler) {
-            $this->isolator()->restore_error_handler();
+            $this->isolator->restore_error_handler();
         }
 
-        $this->isolator()->restore_error_handler();
+        $this->isolator->restore_error_handler();
 
         return $handler;
     }